fix(experimental): store display coordinates as float32, not rounded integers
Rounding to whole pixels put every transcript on an integer lattice. The
displacement is scientifically negligible -- mean 0.08 um against Xenium's
~0.1-0.3 um localisation precision -- but visually obvious, because regular
quantisation creates structure the eye reads as real.

float32 was chosen over fixed-point integers despite storing less densely. On
Xenium pancreas the render file measures:

  whole pixels (uint32)        27.6 MB   1 px      max err 0.150 um
  fixed-point x4 (uint32)      43.3 MB   1/4 px    max err 0.038 um
  float32                      78.5 MB   ~0.004 px max err 0

Fixed-point is denser, but requires the client to apply a scale read from the
manifest; a client that ignores it renders everything silently offset by that
factor. float32 needs no client-side arithmetic and cannot be misread. The extra
~50 MB is about 1% of a 3.5 GB store, which is a good trade for removing a whole
class of silent-wrongness.

Applies to display_geometry as well, so polygon vertices are no longer quantised
either, and deck.gl still receives the Arrow buffer with no copy.

+++ tests/test_shapes_parquet.py
@@ -150,7 +150,7 @@ def test_display_geometry_has_the_nested_layout(rendered: tuple[Path, dict]) -> 
     t = pq.ParquetFile(out).schema_arrow.field(GEOMETRY_COLUMN).type
     assert pa.types.is_list(t)  # polygon level
     assert pa.types.is_list(t.value_type)  # ring level
-    assert t.value_type.value_type == pa.list_(pa.uint32(), 2)  # interleaved vertices
+    assert t.value_type.value_type == pa.list_(pa.float32(), 2)  # interleaved vertices
 
 
 def test_display_geometry_offsets_resolve_like_the_js_reader(rendered: tuple[Path, dict]) -> None:
